requirements.txt
pyspf
pydns